fix: rename DTO -> Models and Models -> Entities
