+++ src/SocialMediaDashboard.Common/Models/AuthenticationResult.cs
@@ -1,4 +1,4 @@
-﻿namespace SocialMediaDashboard.Common.DTO
+﻿namespace SocialMediaDashboard.Common.Models
 {
     /// <summary>
     /// Authentication data transfet object.
