+++ src/SocialMediaDashboard.Common/Models/UserResult.cs
@@ -1,4 +1,4 @@
-﻿namespace SocialMediaDashboard.Common.DTO
+﻿namespace SocialMediaDashboard.Common.Models
 {
     /// <summary>
     /// User data transfet object.
