+++ src/SocialMediaDashboard.Data/Configurations/MediaConfiguration.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using SocialMediaDashboard.Domain.Models;
+using SocialMediaDashboard.Domain.Entities;
 using System;
 
 namespace SocialMediaDashboard.Data.Configurations
