+++ src/SocialMediaDashboard.Domain/Entities/Profile.cs
@@ -2,7 +2,7 @@
 using SocialMediaDashboard.Common.Interfaces;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace SocialMediaDashboard.Domain.Models
+namespace SocialMediaDashboard.Domain.Entities
 {
     /// <summary>
     /// User profile.
