+++ src/SocialMediaDashboard.Domain/Entities/RefreshToken.cs
@@ -3,7 +3,7 @@ using SocialMediaDashboard.Common.Interfaces;
 using System;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace SocialMediaDashboard.Domain.Models
+namespace SocialMediaDashboard.Domain.Entities
 {
     /// <summary>
     /// Refresh Token.
